pyproject.toml
[project]
name = "mira-project"
version = "0.1.0"
description = "mira sdk projec"
authors = [
    {name = "dannyboy-1412",email = "danielantorodri@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.6,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pymongo[srv] (>=4.10.1,<5.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "openai (>=1.60.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
